fix(code): prevent billing plan flicker on reload
On reload, /dashboard/billing flashed the plan chooser before the
active-plan billing UI because the dev-plan status query is gated on
the user query (enabled: !!user), so it stays disabled (and undefined)
on the first client render.

Server-render the user + dev-plan status in the (main) layout and the
payment method in the billing page, feeding them as React Query
initialData so the correct shell renders on first paint. Add a Skeleton
component and replace the spinner / null returns with skeleton loading
states as a genuine-loading fallback.
