+++ apps/code/src/app/dashboard/(main)/layout.tsx
@@ -3,19 +3,26 @@ import { redirect } from "next/navigation";
 import DashboardShell from "@/app/dashboard/DashboardShell";
 import { fetchServerData } from "@/lib/server-api";
 
+import type { DevPlanStatus } from "@/app/dashboard/useDevPlanStatus";
+import type { UserMe } from "@/hooks/useUser";
+
 export default async function DashboardLayout({
 	children,
 }: {
 	children: React.ReactNode;
 }) {
-	const userData = await fetchServerData<{ user: { id: string } } | null>(
-		"GET",
-		"/user/me",
-	);
+	const [userData, devPlanStatus] = await Promise.all([
+		fetchServerData<UserMe>("GET", "/user/me"),
+		fetchServerData<DevPlanStatus>("GET", "/dev-plans/status"),
+	]);
 
 	if (!userData?.user) {
 		redirect("/login?returnUrl=/dashboard");
 	}
 
-	return <DashboardShell>{children}</DashboardShell>;
+	return (
+		<DashboardShell initialUser={userData} initialDevPlanStatus={devPlanStatus}>
+			{children}
+		</DashboardShell>
+	);
 }
