+++ apps/code/src/hooks/useUser.ts
@@ -7,9 +7,15 @@ import { useEffect } from "react";
 import { useAppConfig } from "@/lib/config";
 import { useApi } from "@/lib/fetch-client";
 
+import type { paths } from "@/lib/api/v1";
+
+export type UserMe =
+	paths["/user/me"]["get"]["responses"]["200"]["content"]["application/json"];
+
 export interface UseUserOptions {
 	redirectTo?: string;
 	redirectWhen?: "authenticated" | "unauthenticated";
+	initialData?: UserMe | null;
 }
 
 export function useUser(options?: UseUserOptions) {
@@ -28,6 +34,7 @@ export function useUser(options?: UseUserOptions) {
 			retry: 0,
 			staleTime: 5 * 60 * 1000,
 			refetchOnWindowFocus: false,
+			initialData: options?.initialData ?? undefined,
 		},
 	);
